Change App layout, add basic mode toggle component
