+++ examples/sandbox/src/Theme.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, ParentComponent, Component, createSignal } from 'solid-js'
+import { createContext, useContext, ParentComponent, Component } from 'solid-js'
 import { createStore } from 'solid-js/store'
 
 export type ThemeContextState = {
