+++ examples/sandbox/src/devtools.ts
@@ -1,0 +1,13 @@
+import 'solid-devtools'
+import { attachDevtoolsOverlay } from '@solid-devtools/overlay'
+import { useLocator } from '@solid-devtools/debugger'
+
+useLocator({
+  targetIDE: 'vscode',
+})
+
+!process.env.EXT &&
+  attachDevtoolsOverlay({
+    defaultOpen: true,
+    noPadding: true,
+  })
