+++ examples/sandbox/src/index.tsx
@@ -1,24 +1,11 @@
 /* @refresh reload */
 import { render } from 'solid-js/web'
 
-import 'solid-devtools'
-import { attachDevtoolsOverlay } from '@solid-devtools/overlay'
-import { useLocator } from '@solid-devtools/debugger'
-
-useLocator({
-  targetIDE: 'vscode',
-})
-
-!process.env.EXT &&
-  attachDevtoolsOverlay({
-    defaultOpen: true,
-    noPadding: true,
-  })
-
 import App from './App'
 import { ThemeProvider } from './Theme'
 
 import 'uno.css'
+import './devtools'
 
 export const disposeApp = render(
   () => (
