+++ packages/frontend/src/App.tsx
@@ -1,71 +1,10 @@
-import { Component, createSignal, JSX, Show } from 'solid-js'
+import { Component, JSX, Show } from 'solid-js'
 import { Menu, MenuItem, Popover, PopoverButton, PopoverPanel } from 'solid-headless'
-import { createShortcut } from '@solid-primitives/keyboard'
-import { Splitter, ToggleButton, Icon } from '@/ui'
+import { Splitter, Icon } from '@/ui'
 import Inspector from './modules/inspector/Inspector'
 import Structure from './modules/structure/Structure'
 import { useController } from './controller'
 import * as styles from './App.css'
-
-const LocatorButton: Component = () => {
-  const ctx = useController()
-  return (
-    <ToggleButton
-      class={styles.locatorButton}
-      onToggle={ctx.setLocatorState}
-      selected={ctx.locatorEnabled()}
-    >
-      <Icon.Select class={styles.locatorIcon} />
-    </ToggleButton>
-  )
-}
-
-const Search: Component = () => {
-  const ctx = useController()
-
-  const [value, setValue] = createSignal('')
-
-  const handleChange = (v: string) => {
-    setValue(v)
-    ctx.searchStructure('')
-  }
-
-  return (
-    <form
-      class={styles.search.form}
-      onSubmit={e => {
-        e.preventDefault()
-        ctx.searchStructure(value())
-      }}
-      onReset={() => handleChange('')}
-    >
-      <input
-        ref={input => {
-          if (ctx.options.useShortcuts) {
-            createShortcut(['/'], () => input.focus())
-            createShortcut(['Escape'], () => {
-              if (document.activeElement === input) input.blur()
-              if (input.value) handleChange((input.value = ''))
-            })
-          }
-        }}
-        class={styles.search.input}
-        type="text"
-        placeholder="Search"
-        onInput={e => handleChange(e.currentTarget.value)}
-        onPaste={e => handleChange(e.currentTarget.value)}
-      />
-      <div class={styles.search.iconContainer}>
-        <Icon.Search class={styles.search.icon} />
-      </div>
-      {value() && (
-        <button class={styles.search.clearButton} type="reset">
-          <Icon.Close class={styles.search.clearIcon} />
-        </button>
-      )}
-    </form>
-  )
-}
 
 const Options: Component = () => {
   return (
@@ -105,8 +44,6 @@ const App: Component<{ headerSubtitle?: JSX.Element }> = props => {
   return (
     <div class={styles.app}>
       <header class={styles.header}>
-        <LocatorButton />
-        <Search />
         <div>
           <h3>Solid Devtools</h3>
           {props.headerSubtitle && <p class={styles.subtitle}>{props.headerSubtitle}</p>}
