+++ packages/frontend/src/ui/theme/index.ts
@@ -1,165 +1,3 @@
-// current theme is based on the tailwind default theme
-// colors: https://github.com/tailwindlabs/tailwindcss/blob/master/src/public/colors.js
-// spacing and other values: https://github.com/tailwindlabs/tailwindcss/blob/master/stubs/defaultConfig.stub.js
-
-const cyan = {
-  50: '#ecfeff',
-  100: '#cffafe',
-  200: '#a5f3fc',
-  300: '#67e8f9',
-  400: '#22d3ee',
-  500: '#06b6d4',
-  600: '#0891b2',
-  700: '#0e7490',
-  800: '#155e75',
-  900: '#164e63',
-} as const
-
-const amber = {
-  50: '#fffbeb',
-  100: '#fef3c7',
-  200: '#fde68a',
-  300: '#fcd34d',
-  400: '#fbbf24',
-  500: '#f59e0b',
-  600: '#d97706',
-  700: '#b45309',
-  800: '#92400e',
-  900: '#78350f',
-} as const
-
-const gray = {
-  50: '#fafafa',
-  100: '#f4f4f5',
-  200: '#e4e4e7',
-  300: '#d4d4d8',
-  400: '#a1a1aa',
-  500: '#71717a',
-  600: '#52525b',
-  700: '#3f3f46',
-  800: '#27272a',
-  900: '#18181b',
-} as const
-
-const fontSize = {
-  xs: '0.5rem',
-  sm: '0.625rem',
-  base: '0.75rem',
-  lg: '0.875rem',
-  xl: '1rem',
-  '2xl': '1.125rem',
-  '3xl': '1.25rem',
-} as const
-
-const lineHeight = {
-  '3xs': '1rem',
-  '2xs': '1rem',
-  xs: '1rem',
-  sm: '1.25rem',
-  base: '1.5rem',
-  lg: '1.75rem',
-  xl: '1.75rem',
-  '2xl': '2rem',
-  '3xl': '2.25rem',
-  '4xl': '2.5rem',
-  '5xl': '1',
-  '6xl': '1',
-} as const
-
-export const theme = {
-  color: {
-    inherit: 'inherit',
-    current: 'currentColor',
-    transparent: 'transparent',
-    black: '#000000',
-    white: '#ffffff',
-    cyan,
-    amber,
-    gray,
-    green: '#16a34a',
-    red: '#ef4444',
-    disabled: {
-      light: gray[500],
-      dark: gray[400],
-    },
-  },
-  spacing: {
-    px: '1px',
-    0: '0',
-    0.25: '0.0625rem',
-    0.5: '0.125rem',
-    1: '0.25rem',
-    1.5: '0.375rem',
-    2: '0.5rem',
-    2.5: '0.625rem',
-    3: '0.75rem',
-    3.5: '0.875rem',
-    4: '1rem',
-    4.5: '1.125rem',
-    5: '1.25rem',
-    6: '1.5rem',
-    7: '1.75rem',
-    8: '2rem',
-    9: '2.25rem',
-    10: '2.5rem',
-    11: '2.75rem',
-    12: '3rem',
-    14: '3.5rem',
-    16: '4rem',
-    20: '5rem',
-    24: '6rem',
-    28: '7rem',
-    32: '8rem',
-    36: '9rem',
-    40: '10rem',
-    44: '11rem',
-    48: '12rem',
-    52: '13rem',
-    56: '14rem',
-    60: '15rem',
-    64: '16rem',
-    72: '18rem',
-    80: '20rem',
-    96: '24rem',
-  },
-  easing: {
-    DEFAULT: 'cubic-bezier(0.4, 0, 0.2, 1)',
-    linear: 'linear',
-    in: 'cubic-bezier(0.4, 0, 1, 1)',
-    out: 'cubic-bezier(0, 0, 0.2, 1)',
-    'in-out': 'cubic-bezier(0.4, 0, 0.2, 1)',
-  },
-  radius: {
-    none: '0px',
-    sm: '0.125rem',
-    DEFAULT: '0.25rem',
-    md: '0.375rem',
-    lg: '0.5rem',
-    xl: '0.75rem',
-    '2xl': '1rem',
-    '3xl': '1.5rem',
-    full: '9999px',
-  },
-  duration: {
-    0: '0ms',
-    75: '75ms',
-    100: '100ms',
-    150: '150ms',
-    200: '200ms',
-    300: '300ms',
-    500: '500ms',
-    700: '700ms',
-    1000: '1000ms',
-  },
-  font: {
-    sans: "ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji'",
-    serif: "ui-serif, Georgia, Cambria, 'Times New Roman', Times, serif",
-    mono: "ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace",
-  },
-  fontSize,
-  lineHeight,
-} as const
-
-export const { spacing, color } = theme
-
+export * from './theme.css'
+export * from './vars.css'
 export * from './utils'
